+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultMenuBuilder.java
@@ -216,7 +216,17 @@ final class DefaultMenuBuilder extends AbstractComponentBuilder<Void, JMenu, Men
 			this.menu = menu;
 			this.menuItemBuilder = menuItemBuilder;
 			this.toggleMenuItemBuilder = toggleMenuItemBuilder;
-			controls.actions().forEach(this);
+			List<Action> actions = trimSeparators(new ArrayList<>(controls.actions()));
+			for (int i = 0; i < actions.size(); i++) {
+				Action action = actions.get(i);
+				if (action == Controls.SEPARATOR) {
+					// Prevent multiple separators
+					if (i > 0 && actions.get(i - 1) == Controls.SEPARATOR) {
+						continue;
+					}
+				}
+				accept(action);
+			}
 		}
 
 		@Override
@@ -244,6 +254,17 @@ final class DefaultMenuBuilder extends AbstractComponentBuilder<Void, JMenu, Men
 		@Override
 		void onAction(Action action) {
 			menu.add(action);
+		}
+
+		private static List<Action> trimSeparators(List<Action> actions) {
+			while (!actions.isEmpty() && actions.get(0) == Controls.SEPARATOR) {
+				actions.remove(0);
+			}
+			while (!actions.isEmpty() && actions.get(actions.size() - 1) == Controls.SEPARATOR) {
+				actions.remove(actions.size() - 1);
+			}
+
+			return actions;
 		}
 	}